refactor with chatGPT
prompt: can you please simplify this code

chatGPT:

Replaced the forEach() method with map(), which returns a new array instead of mutating the original one.
Moved the check for whether a chain has been deployed to the top of the map() function, so that we don't create a row for a chain that doesn't have any deployments.
Replaced the versions.forEach() loop with versions.map(), which returns an array of versions instead of concatenating them into a string.
Removed the console.log() statements.
Chained the filter() and map() methods together to make the code more concise.
Used the Boolean constructor as a shorthand for filtering out empty strings from the array of chain rows.

prompt: can you also use the versions array to create the seperatorRow

chatGPT:

In the headerRow, I used padEnd() to add spaces to the version strings to make them fit nicely in the table.

In the seperatorRow, I used map() to create an array of separator strings for each version and then join() to concatenate them together with a | separator.