Add Code Hashes To All Deployments (#654)
This PR adds code hashes to all deployment files. This is useful for a
couple of reasons:

1. It makes it easier for external parties to verify the validity of the
code at a specific address (even if it isn't at the expected `CREATE2`
address)
2. It simplifies the `github-review.sh` script that @remedcu is working
on in #649
3. Isn't "tech debt" like the `defaultAddress` field, as the code hash
should be the same regardless of the deployment address (so it does not
interfere with the plans to improve the deployment JSON format).

For posterity, the changes were generated with:

```fish
for f in src/assets/*/*.json
  set json "$(jq --arg h (jsonrpc $rpc eth_getCode (jq -r .defaultAddress $f) latest | hdwallet hex decode - | hdwallet hash data -) '{ defaultAddress, released, contractName, version, codeHash: $h, networkAddresses, abi }' $f)"
  echo "$json" > $f
end
```

Note that this makes a few formatting changes to some of the deployment
files, but IMO this is a good thing (as we now have consistent
formatting across all files :tada:).