Add A New GitHub Release Action (#666)
This PR adds a new GitHub workflow action to draft releases that can be
used for publishing to NPM. Namely it will:

- Skip the workflow if the latest released NPM version matches the
current release tag
- Bump the package version if needed and push to GitHub automatically
(the configured username and email were taken from
<https://github.com/actions/checkout/?tab=readme-ov-file#push-a-commit-using-the-built-in-token>)
- Draft a new release (note that this DOES NOT create a new tag, which
is nice so we don't need to worry about force pushing/pulling tags).
- Create an NPM package tarball and attach it to the draft release.

The ultimate goal is that a separate runner will:
- Download the tarball
- Publish it to NPM

### Future Work

Somethings that are **NOT** included in this PR:
- Adding a scheduled run for the release process (this will come later
when we are satisfied with the release process that we have)
- Publishing the GitHub release. Note that this can be trivially done
with:
    ```sh
    gh release edit "$tag" --draft=false
    ```

In order to achieve these things, we plan on using the NPM registry as a
source of truth - if there is an NPM package where the version matches
the latest draft release, then we publish it. This is not included in
the PR as it adds additional complexity and possible states that need to
be handled by the script, and it was already complicated enough for now.
Once we have our publishing runner up and working, then I will add
support for this here.