Feature: Improve the JSON file structure to support multiple addresses per network (#668)
This PR implements
https://github.com/safe-global/safe-deployments/issues/348. The
following has been done:
1. The JSON assets have been updated to the new format to support
multiple addresses. Now, the format is the following:
- **released**: `boolean`
  - Indicates if the contract is released.
- **contractName**: `string`
  - The name of the contract.
- **version**: `string`
  - The version of the contract.
- **codeHash**: `string`
  - The hash of the contract code.
- **addresses**: `object`
  - Contains the canonical addresses for the contract.
  - **type**: `address`
    - The canonical address of the contract.
- **networkAddresses**: `object`
  - Maps network IDs to the canonical addresses.
  - **ID**: `string` or `string[]`
- **abi**: `array`
- The ABI (Application Binary Interface) of the contract, containing
multiple function and event definitions.
2. The backward compatibility is implemented through JSON to JS object
mapping through JS functions. As of now, the functions return the same
structure as previous versions, except for the `addresses` object
addition.
3. To ensure 2), I copied all the old JSON assets to the tests folder
and compared them against the newly returned objects

**Important:**
This PR doesn't add any new functions to make use of the new JSON
structure. I decided to bring this out for review and temperature check
and if this. gets accepted, I'll implement the functions in a new PR