Readme updates and improvements to the deployment process (#678)
This PR proposes README updates after
https://github.com/safe-global/safe-deployments/issues/348, namely:

1. Add more detailed steps to add new deployments to the package
2. Grammar fixes
3. Describe the meaning of returned values
4. Remove the obsolete issue template for the new chain. I removed it
because we unnecessarily make people open an issue and a PR
simultaneously. We do not deploy the contracts ourselves, so I don't
think we need an issue template. If the latter is not the case, I can
look into switching the template to be more
`safe-singleton-factory`-alike.
5. Remove the migration script I forgot to remove in my previous PR.