Deployment Review Script (#683)
This PR improves the review cycle for the `safe-deployment` by checking
the new deployments when a new PR is created for a particular chain.

This PR:
- instead of an `ISSUE_TEMPLATE`, introduced a PR Template which will
ask for the required data from the PR creator, and help run the CI
script.
- adds a CI review script, which succeeds if everything seems ok, else,
it helps the PR creator to know what might be a problem with merging the
PR in the CI logs.
- updates the `github-review` script, which checks for edge cases when a
chain ID with the highest number is added.
- updates the test to accommodate the change in the JSON structure for
the `codehash` (because `v.1.3.0` has multiple deployment types).
- updated the JSON structure to include address and code hash within a
single object with the deployment type as the key.
- ZKSync support in CI mentioned in #662 (Ex:
https://github.com/remedcu/safe-deployments/pull/8)

Remaining TODO's:
- There is a possibility that a user removes one of the deployment types
(currently only applicable with `v1.3.0`) from the current list. This
can't be detected with the current script.