Refactor GitHub Review Script (#698)
This PR splits the GitHub review script into two pieces:
1. The old review script, that takes a PR and verifies it meets all the
requirements for deployment PRs
2. A new Node.JS script that verifies the actual deployment and code,
but does not check all the PR rules.

The reason for this split is to accomodate PRs like #695, where the RPC
url is not included in PR description, and more than one version is
added at a time. It essentially adds a "manual" network verification
scheme that can be used in special cases the PR above.

Additionally, it removes the requiment on `cast`. This is particularly
important for automated checks, as we do not want to depend on nightly
tools that don't have releases with somewhat guaranteed behaviour.

Lastly, this PR does a couple minor fixes to the review script, in
particular it causes it to fail if more than one version is changed in
the PR, and doesn't print usage information on general errors.