Run on PR Branch (#700)
This PR changes the review script to run on the PR branch in CI instead
of trying to apply changes to the branch. This is done to avoid issues
where merges to `main` cause PRs to no longer review properly.

In the past, we made these changes in order to ensure that the script
gets run as-expected from `main` to make sure that PRs couldn't change
the scripts and cause them to pass, but this is now ensured by the
GitHub action itself.