Fix Release Script Draft Creation
I noticed when inspecting the GitHub release script that it was no
longer creating new draft releases if

This is undesirable behaviour, as it does not let us accumulate
changes for a draft release while the publish is pending. If we switch
to running the draft GitHub action more often, then this will become an
issue.

Additionally, I refactored the script as there was too much
rightward-drift for my liking.