RPC Taken from `ethereum-lists` directly (#873)
Solved #688 

This pull request includes a small but significant change to the
`.github/PULL_REQUEST_TEMPLATE.md` file. The change updates the
instructions for adding a new chain by specifying that the RPC URL will
be taken from the `ethereum-lists/chains` repository.

*
[`.github/PULL_REQUEST_TEMPLATE.md`](diffhunk://#diff-18813c86948efc57e661623d7ba48ff94325c9b5421ec9177f724922dd553a35L3-L12):
Updated the instructions to clarify that the RPC URL will be sourced
from the `ethereum-lists/chains` repository.
* `bin/github-review.sh`: Updated how RPC is fetched (Before: from PR
description. After: from `ethereum-lists`).