Added Polygon Amoy deployments v1.4.1 Chain ID: 80002 (#1132)
## Add new chain

> Template to provide information about the new chain. Only add extra
information in the bottom section. Ensure that the contracts are
deployed on the chain, if not deploy with
[safe-contracts](https://github.com/safe-global/safe-contracts). Only
**one** chain ID, **one** Safe version and **one** deployment type per
PR. The RPC will be taken from
[ethereum-lists/chains](https://github.com/ethereum-lists/chains). This
entire paragraph should be deleted.

Please fill the following form:

Provide the Chain ID (Only 1 chain id per PR).
- Chain_ID: 80002

Relevant information:
Seems this was an omission. Deployments were done over a year ago