Network Review CI Improvements (#1143)
Some small improvements to the review CI workflow:

* Bump the Node JS version to v22 - this is the current LTS version
(this also applies to other workflows)
* Make sure the diff covers _all_ files that are not assets. This is
important, as it avoids changes to the workflow file, but also the
`scripts/review/verifyDeployment.js` file
* Split out the review verification and dependency installation to
separate steps; for `npm ci` this has a nice security benefit that the
`GH_TOKEN` is not available when installing packages. It also makes it
easier to diagnose exactly what step fails in CI - (see #1142, where the
CI output isn't clear if `npm ci` is exiting with code 22, or the review
script is).