Use DeFi Llama Chainlist (#1158)
We require that the chain appears on https://chainlist.org/, and
previously, this meant that the chain was included in the
ethereum-lists/chains repository. However, since that repository is no
longer actively maintained, and since then DeFi Llama has added
mechanisms for "additional chains" to their DeFiLlama/chainlist
repository, which is the code behind https://chainlist.org/.

This PR changes our scripts to use the Chainlist.org API for checking if
a network is listed or not (instead of the ethereum-lists/chains
repository).

---------

Co-authored-by: Nicholas Rodrigues Lordello <nick@safe.global>