Fix Asset Ordering Test
This PR fixes the asset ordering test. The splitting on `,` was
causing the "canonical" and "eip155" to be parsed as keys, which was
turning into `NaN` and causing sorting to be unstable leading to false
negatives (i.e. test would pass despite keys not being ordered).

The fix is to split on newlines, and check that only one object key
value pair exists per line.