LENS 1.3.0 (#1222)
## Add new chain

> Template to provide information about the new chain. Only add extra
information in the bottom section. Ensure that the contracts are
deployed on the chain, if not deploy with
[safe-contracts](https://github.com/safe-global/safe-contracts). Only
**one** chain ID, **one** Safe version and **one** deployment type per
PR. The RPC will be taken from [DefiLlama's
ChainList](https://chainlist.org/). This entire paragraph should be
deleted.

Please fill the following form:

Provide the Chain ID (Only 1 chain id per PR).

- Chain_ID: 232

Relevant information:
https://explorer.lens.xyz/